Use full URL for filename, not just hostname.
This allows for separate filenames for http://foo.com/,
https://foo.com/, https://joe:user@foo.com/, http://foo.com/path1/,
http://foo.com/path2/.  I wanted to be able to compare /asdf/ from
a bunch of servers, or compare screenshots when different basic-auth
credentials had been used when fetching.

This implementation simply URL-quotes the entire URL, which results
in a messy, but filesystem-safe name.  It might be desirable to be
a bit more selective to make the filename less ugly.

Should be both python2 and python3 safe.